Subsume RelayRequest into relay_handler module
